refactored code
