+++ exploration_benchmark.py
@@ -8,12 +8,12 @@ from agents.exploration_agent import PPOAgent
 from config import get_config
 
 from config import get_config, MyConfigPlugin
-from habitat_extensions.habitat_lab.tasks.exp.exp import ExplorationTask  # noqa (register ExplorationTask)
+from habitat_extensions.habitat_lab.tasks.exp.task import ExplorationTask  # noqa (register ExplorationTask)
 
 # registrations
 import config.default_structured_configs  # noqa structured configs
-import habitat_extensions.habitat_lab.tasks.exp.exp # noqa ExplorationVisitedLocationsReward
-import habitat_extensions.habitat_lab.tasks.nav.nav # noqa TopDownMap
+import habitat_extensions.habitat_lab.tasks.exp.task # noqa ExplorationVisitedLocationsReward
+import habitat_extensions.habitat_lab.tasks.nav.measures # noqa TopDownMap
 import habitat_extensions.habitat_lab.datasets.exploration_dataset # noqa register Exploration datasets
